Cambio tutti i throw in responseEntity

+++ src/main/java/com/example/EasyBlog/Repositories/ArticlesRepository.java
@@ -2,8 +2,6 @@ package com.example.EasyBlog.Repositories;
 
 import com.example.EasyBlog.Entity.Articles;
 import com.example.EasyBlog.Entity.Enum.TypeGenderArticlesEnum;
-import com.example.EasyBlog.Entity.Enum.TypeStatusEnum;
-import com.example.EasyBlog.Entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -13,7 +11,7 @@ import java.util.Optional;
 public interface ArticlesRepository extends JpaRepository<Articles,Long> {
 
     @Query("UPDATE Articles SET typeStatus = ? WHERE id = ? ")
-    Optional<Articles> updateArticlesStatus(Integer articleId);
+    Optional<Articles> updateArticlesStatus(Long articleId);
 
     @Query("SELECT a FROM Articles a WHERE a.typeGender = ?1")
     Optional<List<Articles>> getArticlesByGender(TypeGenderArticlesEnum typeGenderArticlesEnum);
